error

+++ ClientSide/src/http.js
@@ -28,7 +28,7 @@ instance.interceptors.response.use(function (response) {
 }, function (error) {
     // Any status codes that falls outside the range of 2xx cause this function to trigger
     // Do something with response error
-    if (error.response.status === 401 || error.response.status === 403) {
+    if (error.response && (error.response.status === 401 || error.response.status === 403)) {
         localStorage.clear();
         window.location = "/login";
     }
